Removed the dependency on the pool from the Entity class.
In the original implementation, the Entity a b c had a dependency
a -> c, where c was the pool used to generate random entities.
However, each user of the library needs to define his own
implementation of the Entity class and it's functions. Hence, they
know in their module what that pool will be, avoiding the need to
pass it along to the evolutionary algortithm. Indeed, the GA itself
need not know that random entities are based on a pool or on anything
besides the requirement for a seed.