Partially fixored the monadic extension to Entity
Added a new class after a hint from applicative on #haskell. The
functions that are used be Entity but are not dependent on the Monad m
are declared in a separate Mutator class. Entity then depends on this
Mutator, such that all datatypes that purport to be an Entity first must
be a Mutator.

Extracted several parts of the evolutionStep function into smaller
functions, that are defined on the toplevel, i.e., to determine the new
entities obtained from crossover and mutation.